+++ android/common/src/main/java/com/marianhello/bgloc/data/sqlite/SQLiteConfigurationDAO.java
@@ -57,8 +57,9 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
       ConfigurationEntry.COLUMN_NAME_HEADERS,
       ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS,
       ConfigurationEntry.COLUMN_NAME_TEMPLATE,
-      // fork: min interval used to post coordinates to the server
-      ConfigurationEntry.COLUMN_NAME_MIN_POST_INTERVAL
+      // fork: min interval used to post coordinates to the server and max location age
+      ConfigurationEntry.COLUMN_NAME_MIN_POST_INTERVAL,
+      ConfigurationEntry.COLUMN_NAME_MAXIMUM_AGE
     };
 
     String whereClause = null;
@@ -125,8 +126,9 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
     config.setHttpHeaders(new JSONObject(c.getString(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_HEADERS))));
     config.setMaxLocations(c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS)));
     config.setTemplate(LocationTemplateFactory.fromJSONString(c.getString(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_TEMPLATE))));
-    // fork: min interval used to post coordinates to the server
+    // fork: min interval used to post coordinates to the server and max location age
     config.setMinPostInterval(c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_MIN_POST_INTERVAL)));
+    config.setMaximumAge(c.getLong(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_MAXIMUM_AGE)));
 
     return config;
   }
@@ -158,8 +160,9 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
     values.put(ConfigurationEntry.COLUMN_NAME_HEADERS, new JSONObject(config.getHttpHeaders()).toString());
     values.put(ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS, config.getMaxLocations());
     values.put(ConfigurationEntry.COLUMN_NAME_TEMPLATE, config.hasTemplate() ? config.getTemplate().toString() : null);
-    // fork: min interval used to post coordinates to the server
+    // fork: min interval used to post coordinates to the server and max location age
     values.put(ConfigurationEntry.COLUMN_NAME_MIN_POST_INTERVAL, config.hasMinPostInterval() ? config.getMinPostInterval() : null);
+    values.put(ConfigurationEntry.COLUMN_NAME_MAXIMUM_AGE, config.hasMaximumAge() ? config.getMaximumAge() : null);
 
     return values;
   }