+++ android/common/src/main/java/com/marianhello/bgloc/data/sqlite/SQLiteOpenHelper.java
@@ -21,8 +21,8 @@ import static com.marianhello.bgloc.data.sqlite.SQLiteLocationContract.LocationE
 public class SQLiteOpenHelper extends android.database.sqlite.SQLiteOpenHelper {
     private static final String TAG = SQLiteOpenHelper.class.getName();
     public static final String SQLITE_DATABASE_NAME = "cordova_bg_geolocation.db";
-    // fork: column COLUMN_NAME_MIN_POST_INTERVAL added
-    public static final int DATABASE_VERSION = 16;
+    // fork: columns COLUMN_NAME_MIN_POST_INTERVAL, COLUMN_NAME_MAXIMUM_AGE added
+    public static final int DATABASE_VERSION = 17;
 
     public static final String TEXT_TYPE = " TEXT";
     public static final String INTEGER_TYPE = " INTEGER";
@@ -117,6 +117,10 @@ public class SQLiteOpenHelper extends android.database.sqlite.SQLiteOpenHelper {
             case 15:
                 alterSql.add("ALTER TABLE " + ConfigurationEntry.TABLE_NAME +
                         " ADD COLUMN " + ConfigurationEntry.COLUMN_NAME_MIN_POST_INTERVAL + (INTEGER_TYPE));
+            // fork: max location age stored as ineger but should be retrieved as long! cursor.getLong
+            case 16:
+                alterSql.add("ALTER TABLE " + ConfigurationEntry.TABLE_NAME +
+                        " ADD COLUMN " + ConfigurationEntry.COLUMN_NAME_MAXIMUM_AGE + (INTEGER_TYPE));
 
                 break; // DO NOT FORGET TO MOVE DOWN BREAK ON DB UPGRADE!!!
             default: