+++ android/common/src/main/java/com/marianhello/bgloc/headless/ActivityTask.java
@@ -1,5 +1,7 @@
 package com.marianhello.bgloc.headless;
 
+import android.os.Bundle;
+
 import com.marianhello.bgloc.data.BackgroundActivity;
 
 import org.json.JSONException;
@@ -17,7 +19,20 @@ public abstract class ActivityTask extends Task {
     }
 
     @Override
-    String getParams() {
+    public Bundle getBundle() {
+        Bundle bundle = new Bundle();
+        Bundle params = new Bundle();
+
+        params.putInt("confidence", mActivity.getConfidence());
+        params.putString("type", BackgroundActivity.getActivityString(mActivity.getType()));
+
+        bundle.putString("name", getName());
+        bundle.putBundle("params", params);
+        return bundle;
+    }
+
+    @Override
+    public String toString() {
         if (mActivity == null) {
             return null;
         }