+++ android/common/src/main/java/com/marianhello/bgloc/service/LocationService.java
@@ -4,11 +4,13 @@ import com.marianhello.bgloc.Config;
 
 public interface LocationService {
     void start();
+    void startForegroundService();
     void stop();
     void startForeground();
     void stopForeground();
     void configure(Config config);
     void registerHeadlessTask(String jsFunction);
     void startHeadlessTask();
+    void stopHeadlessTask();
     void executeProviderCommand(int command, int arg);
 }