+++ android/common/src/main/java/com/marianhello/bgloc/service/LocationServiceProxy.java
@@ -1,8 +1,8 @@
 package com.marianhello.bgloc.service;
 
-import android.app.ActivityManager;
 import android.content.Context;
 import android.content.Intent;
+import android.os.Build;
 
 import com.marianhello.bgloc.Config;
 
@@ -31,17 +31,29 @@ public class LocationServiceProxy implements LocationService, LocationServiceInf
     }
 
     @Override
-    public void registerHeadlessTask(String jsFunction) {
+    public void registerHeadlessTask(String taskRunnerClass) {
         Intent intent = mIntentBuilder
-                .setCommand(CommandId.REGISTER_HEADLESS_TASK, jsFunction)
+                .setCommand(CommandId.REGISTER_HEADLESS_TASK, taskRunnerClass)
                 .build();
         executeIntentCommand(intent);
     }
 
     @Override
     public void startHeadlessTask() {
+        if (!isStarted()) { return; }
+
         Intent intent = mIntentBuilder
                 .setCommand(CommandId.START_HEADLESS_TASK)
+                .build();
+        executeIntentCommand(intent);
+    }
+
+    @Override
+    public void stopHeadlessTask() {
+        if (!isStarted()) { return; }
+
+        Intent intent = mIntentBuilder
+                .setCommand(CommandId.STOP_HEADLESS_TASK)
                 .build();
         executeIntentCommand(intent);
     }
@@ -57,6 +69,16 @@ public class LocationServiceProxy implements LocationService, LocationServiceInf
 //        intent.addFlags(Intent.FLAG_FROM_BACKGROUND);
         // start service to keep service running even if no clients are bound to it
         executeIntentCommand(intent);
+    }
+
+    @Override
+    public void startForegroundService() {
+        Intent intent = mIntentBuilder.setCommand(CommandId.START_FOREGROUND_SERVICE).build();
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            mContext.startForegroundService(intent);
+        } else {
+            mContext.startService(intent);
+        }
     }
 
     @Override