+++ android/common/src/main/java/com/marianhello/bgloc/sync/NotificationHelper.java
@@ -1,4 +1,4 @@
-package com.marianhello.bgloc;
+package com.marianhello.bgloc.sync;
 
 import android.app.Notification;
 import android.app.NotificationChannel;
@@ -12,6 +12,7 @@ import android.os.Build;
 import android.support.annotation.RequiresApi;
 import android.support.v4.app.NotificationCompat;
 
+import com.marianhello.bgloc.ResourceResolver;
 import com.marianhello.logging.LoggerManager;
 
 public class NotificationHelper {
@@ -20,10 +21,8 @@ public class NotificationHelper {
     public static final String ANDROID_PERMISSIONS_CHANNEL_ID = "android-permissions";
 
     public static final String SYNC_CHANNEL_ID = "syncservice";
-    // fork: service name moved to resources
-    // public static final String SYNC_CHANNEL_NAME = "Sync Service";
-    // fork: not used
-    // public static final String SYNC_CHANNEL_DESCRIPTION = "Shows sync progress";
+    public static final String SYNC_CHANNEL_NAME = "Sync Service";
+    public static final String SYNC_CHANNEL_DESCRIPTION = "Shows sync progress";
 
     public static class NotificationFactory {
         private Context mContext;
@@ -88,27 +87,23 @@ public class NotificationHelper {
 
     public static void registerAllChannels(Context context) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            // fork: channel names changed using resources
-            String serviceChannelName = ResourceResolver.newInstance(context).getString(("mauron85_bgloc_service_name"));
-            String syncChannelName = ResourceResolver.newInstance(context).getString(("mauron85_bgloc_sync_service_name"));
-            String permChannelName = ResourceResolver.newInstance(context).getString(("mauron85_bgloc_perm_service_name"));
+            String appName = ResourceResolver.newInstance(context).getString(("app_name"));
             // Create the NotificationChannel, but only on API 26+ because
             // the NotificationChannel class is new and not in the support library
             android.app.NotificationManager notificationManager = (android.app.NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
-            notificationManager.createNotificationChannel(createServiceChannel(serviceChannelName));
-            notificationManager.createNotificationChannel(createSyncChannel(syncChannelName));
-            notificationManager.createNotificationChannel(createAndroidPermissionsChannel(permChannelName));
+            notificationManager.createNotificationChannel(createServiceChannel(appName));
+            notificationManager.createNotificationChannel(createSyncChannel());
+            notificationManager.createNotificationChannel(createAndroidPermissionsChannel(appName));
         }
     }
 
     public static void registerServiceChannel(Context context) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            // fork: service name channel changed using resources
-            String serviceChannelName = ResourceResolver.newInstance(context).getString(("mauron85_bgloc_service_name"));
+            String appName = ResourceResolver.newInstance(context).getString(("app_name"));
             // Create the NotificationChannel, but only on API 26+ because
             // the NotificationChannel class is new and not in the support library
             android.app.NotificationManager notificationManager = (android.app.NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
-            notificationManager.createNotificationChannel(createServiceChannel(serviceChannelName));
+            notificationManager.createNotificationChannel(createServiceChannel(appName));
         }
     }
 
@@ -116,10 +111,8 @@ public class NotificationHelper {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             // Create the NotificationChannel, but only on API 26+ because
             // the NotificationChannel class is new and not in the support library
-            // fork: service name channel changed using resources
-            String syncChannelName = ResourceResolver.newInstance(context).getString(("mauron85_bgloc_sync_service_name"));
             android.app.NotificationManager notificationManager = (android.app.NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
-            notificationManager.createNotificationChannel(createSyncChannel(syncChannelName));
+            notificationManager.createNotificationChannel(createSyncChannel());
         }
     }
 
@@ -127,19 +120,14 @@ public class NotificationHelper {
     public static NotificationChannel createServiceChannel(CharSequence name) {
         NotificationChannel channel = new NotificationChannel(SERVICE_CHANNEL_ID, name, android.app.NotificationManager.IMPORTANCE_LOW);
         channel.enableVibration(false);
-        // fork: do not show badge counter
-        channel.setShowBadge(false);
         return channel;
     }
 
     @RequiresApi(api = Build.VERSION_CODES.O)
-    public static NotificationChannel createSyncChannel(CharSequence name){
-        NotificationChannel channel = new NotificationChannel(SYNC_CHANNEL_ID, name, NotificationManager.IMPORTANCE_LOW);
-        // fork: escluded because we really do not need syncronization
-        // channel.setDescription(SYNC_CHANNEL_DESCRIPTION);
+    public static NotificationChannel createSyncChannel(){
+        NotificationChannel channel = new NotificationChannel(SYNC_CHANNEL_ID, SYNC_CHANNEL_NAME, NotificationManager.IMPORTANCE_LOW);
+        channel.setDescription(SYNC_CHANNEL_DESCRIPTION);
         channel.enableVibration(false);
-        // fork: do not show badge counter
-        channel.setShowBadge(false);
         return channel;
     }
 
@@ -147,8 +135,6 @@ public class NotificationHelper {
     public static NotificationChannel createAndroidPermissionsChannel(CharSequence name ){
         NotificationChannel channel = new NotificationChannel(ANDROID_PERMISSIONS_CHANNEL_ID, name, NotificationManager.IMPORTANCE_HIGH);
         channel.enableVibration(false);
-        // fork: do not show badge counter
-        channel.setShowBadge(false);
         return channel;
     }
 }